break out tags creation as a function
